+++ django_rubble/utils/test_numeric_utils.py
@@ -3,7 +3,7 @@ from typing import TypeVar
 
 import pytest
 
-from django_rubble.utils.numbers import (
+from django_rubble.utils.numeric_utils import (
     any_to_float,
     is_number,
     ratio_to_whole,
@@ -23,7 +23,7 @@ from django_rubble.utils.numbers import (
         (None, False),
     ],
 )
-def test_is_number(number, expected):
+def test_is_number(number: int | float | str | None, expected: bool):
     assert is_number(number) == expected
 
 
@@ -36,7 +36,7 @@ def test_is_number(number, expected):
         ("3.1400", Decimal("3.14")),
     ],
 )
-def test_trim_trailing_zeros(value, expected):
+def test_trim_trailing_zeros(value: int | float | str, expected: Decimal):
     assert trim_trailing_zeros(value) == expected
     assert Decimal(str(value)).normalize() == expected
 
@@ -49,17 +49,17 @@ def test_trim_trailing_zeros(value, expected):
         ("5.5", 2.3, 5.5),
     ],
 )
-def test_value_as_float(value, default, expected):
+def test_value_as_float(value: str | float, default: float, expected: float):
     assert any_to_float(value, default) == expected
 
 
 def test_any_to_float_exception():
     with pytest.raises(TypeError) as excinfo:
-        any_to_float(4.2, "test")
+        _ = any_to_float(4.2, "test")  # pyright: ignore[reportArgumentType]
     assert str(excinfo.value) == "Default must be of type `float` [test]"
 
 
-T = TypeVar("T", int, float, Decimal)
+_T = TypeVar("_T", int, float, Decimal)
 
 
 @pytest.mark.parametrize(
@@ -70,7 +70,7 @@ T = TypeVar("T", int, float, Decimal)
         (1, 100),
     ],
 )
-def test_ratio_to_whole(input_value: T, expected: T):
+def test_ratio_to_whole(input_value: float | Decimal, expected: float | Decimal):
     assert ratio_to_whole(input_value) == expected
 
 
@@ -82,5 +82,5 @@ def test_ratio_to_whole(input_value: T, expected: T):
         (Decimal("99"), Decimal("0.99")),
     ],
 )
-def test_whole_to_ratio(input_value: T, expected: T):
+def test_whole_to_ratio(input_value: _T, expected: _T):
     assert whole_to_ratio(input_value) == expected